+++ ContentPatcher/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## 1.5.3
+* Added `patch summary` hint if `Target` value incorrectly includes a file extension.
 * Fixed yet another error setting `EditData` entries to `null` since 1.5.
 
 ## 1.5.2